Changes to RobotsTxtServer
- Improved robots.txt handling: when a wildcard rule disallows all and a
specific rule allows a user-agent, the wildcard rule was applied
nonetheless. The change is to maintain two sets: one of generic directives and one
specific for the user agent. If either of the two allows a page, it is
accepted.

- Fixed a bug where the Robots.txt would return a valid HTTP response but
with content length 0. After this fix, the content will be ignored and
treated as an absent robots.txt.

- Added setting ignoreUserAgentInAllow which specifies if the user-agent
should be ignored in allow rules. This basically makes all allow rules
apply to all user agents. However, they can still be overridden by a
Disallow-rule that specifically targets the configured user agent. This
is to overcome weird webmasters that allow some robots and disallow all
other robots.

- Changed decision logic: a URL is now not considered allowed when it is
either allowed and not disallowed by either the specific and the
wildcard user agents. Instead, when a User-agent is specifically
disallowed access, this cannot be overriden anymore by the more general
case.

- Added a configuration option to ignore robots.txt when that file
disallows a seed URL. When setIgnoreRobotsTxtForSeed is set to true, any
seed URL will be crawled regardless of robots.txt configuration.
However, should this page redirect, the robots.txt on the new location
will still be respected.