Fix bug that allows a URL in progress to be re-selected right away, even before it's finished by the current crawler, if it's re-scheduled.
This could happen because in_progress_db and crawl_queue_db are
separate. If the client code re-enqueues a URL during crawling (for
example after a time-out), it will be added to crawl_queue_db while it's
still in in_progress_db. The thread processing it and another thread
waiting for a new URL will then race access to the queue: if the new
thread gets assigned the re-enqueued URL before the old thread marks it
as finished, both will be assigned the same URL but the in_progress_db
will only continue it once due to de-duplication. This will result in a
premature invocation of WebCrawler#handleSeedEnd.